this is our final project

+++ shopping/views.py
@@ -18,9 +18,7 @@ def add_to_cart(request):
         data = json.loads(json_data)
         book_id = data['book_id']
         user_id = data['user_id']
-        # 判断Book数据库里有没有这本书
         book_in_Book = get_object_or_404(Book, id=book_id)
-        # 判断Cart数据库里有没有这本书
         book_in_Cart = Cart.objects.filter(Q(book=book_in_Book), Q(user_id=user_id))
         if book_in_Cart:
             this_user_cart = get_object_or_404(Cart, book=book_in_Book, user_id=user_id)
@@ -28,44 +26,91 @@ def add_to_cart(request):
             this_user_cart.save()
         else:
             Cart(user_id=user_id, book=book_in_Book, quantity=1).save()
-        return redirect('view_cart')
+        return JsonResponse({'message': 'success!'})
+
 
 @csrf_exempt
-def update_cart(request):
+def plusone_cart(request):
     if request.method == 'POST':
         json_data = request.body.decode('utf-8')
         data = json.loads(json_data)
-        action = data['action']
         cart_id = data['cart_id']
-
         cart_item = get_object_or_404(Cart, id=cart_id)
 
-        if action == 'add':
+        if cart_item:
             cart_item.quantity += 1
-        elif action == 'remove':
-            cart_item.quantity -= 1
-
-        if cart_item.quantity < 1:
-            cart_item.delete()
-        else:
             cart_item.save()
-        return redirect('view_cart')
+            return JsonResponse({'message': 'success!'})
+        else:
+            return JsonResponse({'error': 'Without this book!'})
+
+@csrf_exempt
+def minusone_cart(request):
+    if request.method == 'POST':
+        json_data = request.body.decode('utf-8')
+        data = json.loads(json_data)
+        cart_id = data['cart_id']
+        cart_item = get_object_or_404(Cart, id=cart_id)
+
+        if cart_item:
+            cart_item.quantity -= 1
+            if cart_item.quantity < 1:
+                cart_item.delete()
+                return JsonResponse({'message': 'success!'})
+            else:
+                cart_item.save()
+                return JsonResponse({'message': 'success!'})
+
+        else:
+            return JsonResponse({'message': 'Without this book!'})
+
 
 @csrf_exempt
 def view_cart(request):
     if request.method == 'POST':
         json_data = request.body.decode('utf-8')
         data = json.loads(json_data)
-        user_id = data['user_id']
-        user_cart = Cart.objects.filter(Q(user_id=user_id))
+        username = data['userName']
+        get_user = User.objects.get(username=username)
+        user_cart = Cart.objects.filter(Q(user_id=get_user.id))
 
         total = 0
+        book_list = {}
         if user_cart:
             for item in user_cart:
                 total += item.book.price * item.quantity
-            return JsonResponse({"total": total})
+                book_list[item.id] = {'title': item.book.title, 'url': item.book.url, 'author': item.book.author,
+                                      'price': round(item.book.price, 1), 'quantity': item.quantity}
+            return JsonResponse({'book_list': book_list, 'total_price': round(total, 1)})
+
         else:
             return JsonResponse({'message': 'The shopping cart is empty, go shopping!'})
+
+@csrf_exempt
+def delete_item(request):
+    if request.method == 'POST':
+        json_data = request.body.decode('utf-8')
+        data = json.loads(json_data)
+        cart_id = data['cart_id']
+        Cart(id=cart_id).delete()
+
+        return JsonResponse({'message': 'delete success!'})
+
+@csrf_exempt
+def text_bar(request):
+    if request.method == 'POST':
+        json_data = request.body.decode('utf-8')
+        data = json.loads(json_data)
+        username = data['username']
+        cart_id = data['cart_id']
+        new_quantity = data['new_quantity']
+
+        user = get_object_or_404(User, username=username)
+        this_cart = Cart.objects.get(id=cart_id)
+        this_cart.quantity = new_quantity
+        this_cart.save()
+
+        return JsonResponse({'message': 'quantity change success!'})
 
 @csrf_exempt
 def user_transfer(request):
@@ -80,14 +125,69 @@ def user_transfer(request):
             return JsonResponse({'error': 'This user does not exist!'})
 
 @csrf_exempt
-def order_record(request):
+def add_order_record(request):
     if request.method == 'POST':
         json_data = request.body.decode('utf-8')
         data = json.loads(json_data)
-        user_id = data['user_id']
-        total_price = data['total']
-        Order(user_id=user_id, total_price=total_price).save()
-        return JsonResponse({'user_id': user_id, 'total_price': total_price})
+        username = data['userName']
+        total_price = data['total_price']
+        shipping_address = data['address']
+        payment = data['paymentMethod']
+        delivery = data['shippingMethod']
+        book_list = data['book_list']
 
+        get_user = User.objects.get(username=username)
 
+        all_str = str()
+        for index in book_list:
+            this_cart = Cart.objects.get(id=index)
+            this_book = Book.objects.get(id=this_cart.book_id)
+            info_str = str(this_book.title) + '**' + str(this_book.author) + '**' + str(this_cart.quantity) + '**' + str(round(float(this_book.price), 1)) + '**' + str(this_book.url)
+            all_str += info_str + '##'
 
+        Order(user_id=get_user.id, list_book=all_str, total_price=total_price, shipping_address=shipping_address, payment=payment, refund='unrefunded', delivery=delivery, status='confirmed').save()
+        # 清除购物车
+        for index in book_list:
+            Cart(id=index).delete()
+        return JsonResponse({'message': 'Pay Success!'})
+@csrf_exempt
+def view_order_record(request):
+    if request.method == 'POST':
+        json_data = request.body.decode('utf-8')
+        data = json.loads(json_data)
+        username = data['username']
+
+        user = get_object_or_404(User, username=username)
+        user_order_exist = Order.objects.filter(user_id=user.id, refund='unrefunded', status='confirmed').exists()
+        if user_order_exist:
+            user_order = Order.objects.filter(user_id=user.id, refund='unrefunded', status='confirmed')
+            order_list = {}
+            for order in user_order:
+                books_in_order = [x.split('**') for x in order.list_book.split('##')]
+                order_list[order.id] = {'id': order.id, 'user_id': username,
+                                        'books_in_order': books_in_order[:-1],
+                                        'total_price': round(float(order.total_price), 1),
+                                        'shipping_address': order.shipping_address,
+                                        'payment': order.payment, 'refund': order.refund,
+                                        'delivery': order.delivery, 'created_at': order.created_at,
+                                        'updated_at': order.updated_at, 'status': order.status
+                                        }
+            return JsonResponse({'order_list': order_list})
+        else:
+            return JsonResponse({'error': 'Order record does not exist!'})
+
+@csrf_exempt
+def order_delete(request):
+    if request.method == 'POST':
+        json_data = request.body.decode('utf-8')
+        data = json.loads(json_data)
+        username = data['username']
+        order_id = data['order_id']
+
+        user = get_object_or_404(User, username=username)
+        this_order = Order.objects.get(id=order_id)
+        this_order.status = 'cancelled'
+        this_order.refund = 'refunded'
+        this_order.save()
+
+        return JsonResponse({'message': 'delete success!'})
